requirement.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.6
drf-yasg==1.20.0